Move events fixture to AppMapTestBase
